fix-test

+++ sycl/test-e2e/Basic/sycl_ext_oneapi_platform_device_index.cpp
@@ -109,7 +109,7 @@ int main() {
 
   // ext_oneapi_device_at_index
   auto ext_oneapi_device_at_index =
-      plt.ext_oneapi_device_at_index(orig_dev_index_within_plt);
+      plt.ext_oneapi_device_at_index(static_cast<size_t>(orig_dev_index_within_plt));
   assert(orig_dev == ext_oneapi_device_at_index &&
          "A copy of the device object which has that index doesn't match the "
          "original device");
